% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics_coarse_fragments.R
\name{coarse_grade_to_alpha}
\alias{coarse_grade_to_alpha}
\title{Map grade to transparency}
\usage{
coarse_grade_to_alpha(grade)
}
\arguments{
\item{grade}{Character: grade descriptor}
}
\value{
Numeric alpha value (0-1)
}
\description{
Map grade to transparency
}
\keyword{internal}
