% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics_coarse_fragments.R
\name{coarse_size_to_marker}
\alias{coarse_size_to_marker}
\title{Map size class to marker size}
\usage{
coarse_size_to_marker(size_class)
}
\arguments{
\item{size_class}{Character: size descriptor}
}
\value{
Numeric marker size
}
\description{
Map size class to marker size
}
\keyword{internal}
