% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics_boundaries.R
\name{layer_boundary_transition_zones}
\alias{layer_boundary_transition_zones}
\title{Layer for rendering transition zones}
\usage{
layer_boundary_transition_zones(zone_data)
}
\arguments{
\item{zone_data}{Data frame from \code{build_boundary_transition_zones()}}
}
\value{
A ggplot2 layer object (or empty list if no zones)
}
\description{
Layer for rendering transition zones
}
\keyword{internal}
