% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics_coarse_fragments.R
\name{normalize_fragment_type}
\alias{normalize_fragment_type}
\title{Normalize fragment type}
\usage{
normalize_fragment_type(fragment_type)
}
\arguments{
\item{fragment_type}{Character: type descriptor}
}
\value{
Character: normalized type or "gravel" as fallback
}
\description{
Normalize fragment type
}
\keyword{internal}
