% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics_coarse_fragments.R
\name{normalized_fragment_color}
\alias{normalized_fragment_color}
\title{Normalize fragment color}
\usage{
normalized_fragment_color(color)
}
\arguments{
\item{color}{Character or NULL: color specification}
}
\value{
Character: valid color or default
}
\description{
Normalize fragment color
}
\keyword{internal}
