% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{read_soil_json}
\alias{read_soil_json}
\title{Read a soil profile from JSON}
\usage{
read_soil_json(path)
}
\arguments{
\item{path}{Input file path.}
}
\value{
A \code{soil_profile} object.
}
\description{
Read a soil profile from JSON
}
