% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soilgraph-package.R
\docType{package}
\name{soilgraph-package}
\alias{soilgraph}
\alias{soilgraph-package}
\title{soilgraph: Soil Profile Visualization from Field Descriptions}
\description{
Build structured soil profile descriptions following a standardised field specification, export them to a versioned JSON schema, and render publication-ready visualisations with irregular coarse-fragment and horizon-boundary engines powered by 'ggplot2'. Follows Soil Taxonomy (USDA) vocabulary for texture, moisture, colour, boundary grade and shape, and coarse fragment properties.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://nriveras.github.io/soilgraph}
  \item \url{https://github.com/nriveras/soilgraph}
  \item Report bugs at \url{https://github.com/nriveras/soilgraph/issues}
}

}
\author{
\strong{Maintainer}: Nicolás Riveras-Muñoz \email{contact@nriveras.com}

}
\keyword{internal}
